+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.6.0] - ?
 
 - add journal entry
+- add option to protect lives during bloodmoon
 
 ## [0.5.1] - 2022-07-23